OETH withdrawal queue (#2062)
Added a withdrawal queue to the OETH Vault which includes requestWithdrawal, claimWithdrawal and addWithdrawalQueueLiquidity.

Any function that transfers WETH to or from the vault modified so the withdrawal queue is prioritised. This includes allocate, mint, redeem, depositToStrategy, withdrawFromStrategy, withdrawAllFromStrategy, withdrawAllFromStrategies and swapCollateral.

redeem will only work if there is enough WETH liquidity in the Vault to satisfy all outstanding withdrawal requests.

There needs to be a minimum of 10 minute between request and claim. Once the liquidity is available in the vault after the 10 minutes, the user can claim.

OETH Vault's allocate changed to use the OETH total supply rather than calculating the vault's total assets when calculating the vault buffer.

To prevent yield theft, mints over the rebaseThreshold (1 OETH) will call the Dripper to collect any streamed rewards to that point. The minter's OETH is calculated after the rewards are collected. This prevents someone minting with a large amount and then withdrawing 10 minutes later collecting most of the yield from the Dripper.


Co-authored-by: Domen Grabec <grabec@gmail.com>
Co-authored-by: Shahul Hameed <10547529+shahthepro@users.noreply.github.com>
Co-authored-by: Clément <55331875+clement-ux@users.noreply.github.com>
Co-authored-by: clement-ux <clemmoller@gmail.com>